# perf: add `shuffle_list` function