**Motivation**
Shuffling a whole list should be much more efficient than calling `compute_shuffled_index` on every single item.

**Description**
Lighthouse implementations claims it can be up to 250x more efficient than computing `compute_shuffled_index` through the entire list, see: https://github.com/sigp/lighthouse/blob/c11638c36c25f3af910eecd33a6e45d1caa99f1f/consensus/swap_or_not_shuffle/src/shuffle_list.rs#L55

In my case I've observed "only" 10x improvements so far:

```
Name                             ips        average  deviation         median         99th %
shuffle_list                  254.79        3.92 ms    ±10.57%        3.90 ms        4.29 ms
compute_shuffled_index         22.14       45.17 ms     ±2.66%       45.24 ms       52.94 ms

Comparison:
shuffle_list                  254.79
compute_shuffled_index         22.14 - 11.51x slower +41.25 ms
```

TODO: Integrate this function into the code.